added if statment

+++ firewall.sh
@@ -8,4 +8,8 @@ sudo ufw allow in on tun0 from any to any
 sudo ufw allow 1714:1764/udp
 sudo ufw allow 1714:1764/tcp
 sudo ufw enable
+
+if [ -e /home/swipe/.conky/fwoff.png ] 
+then
 mv /home/swipe/.conky/fwoff.png /home/swipe/.conky/fwon.png
+fi
