+++ unfirewall.sh
@@ -11,4 +11,8 @@ sudo ufw allow in 137,138/tcp
 sudo ufw allow in 5556/tcp
 sudo ufw allow in 5558/tcp
 sudo ufw enable
+
+if [ -e /home/swipe/.conky/fwon.png ] 
+then
 mv /home/swipe/.conky/fwon.png /home/swipe/.conky/fwoff.png
+fi
